moving api call to home routes
