+++ public/javascript/movies.js
@@ -1,12 +1,16 @@
 async function getMovies() {
-    const response = await fetch('/api/movie', {
-        method: 'get',
-        body: JSON.stringify({
-            title,
-            poster
-        }),
-        headers: { 'Content-Type': 'application/json' }
-    })
+    const response = await fetch('/api/movie')
+    let data = await response.json()
+    
+    console.log(data)
+    // {
+    //     method: 'get',
+    //     body: JSON.stringify({
+    //         title: title,
+    //         poster: poster
+    //     }),
+        //headers: { 'Content-Type': 'application/json' }
+    // })
 
     console.log(response)
         //check response status
