+++ views/homepage.handlebars
@@ -1,10 +1,10 @@
  <div>You are on the homepage</div>
 
  <div>
-     <img src="{{movie.poster}}">
-     <h2>{{movie.title}}</h2>
+     <img src="{{movies.poster}}">
+     <h2>{{movies.title}}</h2>
  </div>
 
- <button id="movies">Click to See the Movies</button>
+ {{!-- <button id="movies">Click to See the Movies</button> --}}
 
- <script src="/javascript/movies.js"></script>
+ {{!-- <script src="/javascript/movies.js"></script> --}}
